requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.11.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
billiard==4.2.4
bleach==6.2.0
bokeh==3.8.0
celery==5.6.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==6.0.2
django-environ==0.12.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fqdn==1.5.1
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kombu==5.6.2
lark==1.3.0
linkify-it-py==2.0.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
maxminddb==2.6.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.2
multidict==6.1.0
narwhals==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
panel==1.8.1
param==2.2.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prometheus_client==0.23.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.1.0
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.23
Pygments==2.19.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.3.0
pytz==2025.2
pyviz_comms==3.0.6
pywin32==309
pywinpty==3.0.0
PyYAML==6.0.3
pyzmq==26.2.1
redis==7.1.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.23.1
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xyzservices==2025.4.0
yarg==0.1.9
yarl==1.18.3
